Cập nhật một số điện thoại chỉ được đặt một lịch hẹn trong một ngày

+++ src/main/java/fpt/edu/vn/gms/entity/Part.java
@@ -51,7 +51,7 @@ public class Part {
     private BigDecimal discountRate;
 
     @Column(name = "quantity_in_stock")
-    private Integer quantityInStock;
+    private Double quantityInStock;
 
     @Column(name = "reserved_quantity")
     private Integer reservedQuantity;
